+++ vike/node/api/context.ts
@@ -1,16 +1,13 @@
-// export { getOperation }
-export { setOperation }
-export { clearOperation }
 export { isVikeCliOrApi }
+export { setContextApiOperation }
+export { clearContextApiOperation }
+// export { getApiOperation }
 
 import type { Operation } from './types.js'
 import { assert, getGlobalObject } from './utils.js'
+const globalObject = getGlobalObject<{ apiOperation?: Operation }>('api/context.ts', {})
 
-const globalObject = getGlobalObject('context.ts', {
-  apiOperation: undefined as Operation | undefined
-})
-
-function getOperation(): Operation {
+function getApiOperation(): Operation {
   assert(globalObject.apiOperation)
   return globalObject.apiOperation
 }
@@ -18,10 +15,10 @@ function isVikeCliOrApi(): boolean {
   // The CLI uses the API
   return !!globalObject.apiOperation
 }
-function setOperation(operation: Operation): void {
+function setContextApiOperation(operation: Operation): void {
   assert(!globalObject.apiOperation)
   globalObject.apiOperation = operation
 }
-function clearOperation(): void {
+function clearContextApiOperation(): void {
   globalObject.apiOperation = undefined
 }
