+++ vike/node/prerender/context.ts
@@ -3,10 +3,7 @@ export { temp_disablePrerenderAutoRun }
 
 import type { VikeConfigGlobal } from '../plugin/plugins/importUserCode/v1-design/getVikeConfig.js'
 import { getGlobalObject } from '../../utils/getGlobalObject.js'
-
-const globalObject = getGlobalObject<{
-  isDisabled?: true
-}>('isPrerenderAutoRunEnabled.ts', {})
+const globalObject = getGlobalObject<{ isDisabled?: true }>('prerender/context.ts', {})
 
 function isPrerenderAutoRunEnabled(vikeConfigGlobal: VikeConfigGlobal) {
   return (
